Backend/requirements.txt
fastapi~=0.128.0
pydantic-settings~=2.12.0
PyJWT~=2.10.1
pydantic~=2.12.5
SQLAlchemy~=2.0.45
pwdlib~=0.3.0